test(client): move live retry check to e2e
